[CA-272150] Fix behaviour of migration wizard when stepping backward
This page did actually already reload a user's selections when we step
back to it, but it had to reload the combobox and hosts first - which is
slow and makes it look like it's lost the selection.

When we leave the page, we no longer clear the combo box. And when we
enter the page backward, and the combobox already has items saved, we no
longer re-populate it (which involves clearing it). The effect for the
user is that there is no wait when moving back to this page, their
chosen options are already selected and nothing is reloaded
automatically.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>