CA-272143: Import wizard stuck in "Waiting for Import template wizard to complete" when importing a default template
When trying to import a default template from xva, xapi notices that it is a default template, finds an existing one on the host and returns that one.
With these changes, XenCenter handles this case and displays an error, instead of waiting for the returned VM to be linked to the import task (which never happens).

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>