CP-17099: Do the double probe for iSCSI
When we scan the LUN to see if there are any existing SRs, we need to do an additional probe if the probe for the desired type didn't find any SRs.
E.g. if  GFS2 is selected and no GFS2 SR was found on the LUN, then we need to also check if there is a LVM over iSCSI SR, and if there is one we will offer to reattach it (as LVM over iSCSI) or format it as GFS2

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>