CA-286882: Shutting down large number of VMs causes XenCenter to hang
Created a new base class for the Notifications pages (Alerts, Updates, Events). This can be used in the future to host more of the common code between these pages. For now, it implements a common mechanism for showing/hiding the page and ensures that the event handlers are deregistered when the page is hidden (i.e. when the user switches to a different view) and registered again when the page becomes visible

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>