CA-298592: Check the licenseStatus is not null when accessing its properties.
There can be cases when we have logged in a session, events are fired
causing the tab pages to be refreshed, however we may abort the connection,
for example, due to brand mismatch. In this case the licenseStatus has not
been initialized, however the refreshing general tab tries to access its
properties resulting to a crash.
Also some method tidy up.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>