Removed explicit call of UpdateToolbars immediately after the first RequestRefreshTreeView.
Due to the delay interval between refreshes, this UpdateTolbars was called
before any root nodes were added to the tree, hence the only new tab to
add was Search, which got selected  (previously this did not happen,
because it was taken for granted that the root node was added by the constructor).
Immediately afterwards the treeview was refreshed, the root node was selected and the
home page was added, but the selected Search tab was erroneously preserved.
Also removed extra RequestRefreshTreeView because the tree will be refreshed,
if required, by the handlers of the connection events.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>