Suggested modifications.
Renamed ColumnEnabled to ColumnChecked.
Disabled EditMode as it was causing problems to toggling the checked state of multiple
items when the focus was in a checkbox cell.
Although there is no explicit guideline, the expected behaviour for toggling multiple items
seems to be check them all if there are any unchecked/indeterminate, otherwise uncheck.
Rearranged methods to make the file more legible.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>