CA-300759: When there is a GFS2 SR in the pool, VMs on local storage should still be allowed to be migrated to the local storage of another host.
When XenCenter checks if a VM can be storage-migrated (including in the same pool), it only checks if migration is possible to one SR (the first SR from the SR list  that supports vdi creation), which could be a GFS2 SR, where migration is not supported. With this fix, we will only consider the SRs which support migration (instead of vdi creation).

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>