CA-302463: Change the action that applies an update to do an async call, so that the calling code notices if the connection has been lost while installing the update
This fixes the bug in the Patching wizard, where the wizard does not show any error and does not notice the disconnection, so it is stuck in the Installing update phase.
The Patching wizard only uses the ApplyUpdateAction class, but I had also changed the ApplyPatchAction class to avoid similar problems in the RPU wizard, where this action is used to install the RPU hotfix during the prechecks.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>