CA-322699: Do not use an intermediate MemoryStream when sanitising the files of a tar archive.
If the tar contains huge files, this causes memory spikes, even OutOfMemoryException.
Instead, copy the file from the input to the output stream in chunks.
For this purpose the sanitisation code was moved to XenCenterLib to avoid
exposing the tar internals to the UI project.
Also, made some minor improvements in error handling, clean up, and logging.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>