CA-320203: Fixed the update wizard failure when installing two cumula… (#2466)
CA-320203: Fixed the update wizard failure when installing two cumulative updates

The issue was that when installing 2 new versions in one go (e.g. two cumulative updates), the installation succeeds on master but it cannot continue on slaves, because it cannot find the first CU update anymore.
The fix is to refresh the update record before applying the update, with the record found in the cache. Also, if there is no update in the cache, try to re-introduce the update.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>