CA-311165: Modified Display Options
- Removed option to show host only vs pool properties on the Search Tab
when selecting a host within a pool. It is out of place and confusing
because the items under the host don't follow suit. The same result can
be achieved using the filters on the Search Tab.
- Added option to disable remembering the last selected Tab per object
as it can be annoying when trying to compare data on the same Tab for
different objects.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>