CA-322790: Fix unhandled exception when the host cannot be reached to execute an API call.
In the CrossPoolMigrateCanMigrateFilter a WebException can be raised if the host cannot be reached to execute the API call that asserts is the VM can be migrated there. With these changes, we catch any exception (not just Failure) and mark the VM as not migratable (because we couldn't assert if it can be migrated to that host).

Also fixed other places where the same error might occur.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>