CA-329399: Improve the error message displayed if the update wizard if a host cannot be evacuated
If a host cannot be evacuated because other hosts in the pool are maintenance mode and pending a reboot following an update installation, then include instructions to reboot those hosts and then press Retry.
Also updated the text for the Manual mode, removing the reference to installing other updates and doing a reboot manually at the end, because that won't be always possible, due to the fact that now we put the hosts in maintenance mode before installing the update.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>