CP-34000: decimal network and disk access rate unit calculations
Though this was shown in units like kB/s and MB/s (decimal units) it was calculated in
binary terms i.e. 1MB/s was treated as 1024kB/s instead of the expected 1000kB/s.

Updated translations for the units to be of the form 'MB/s' instead of 'MBps'.

Signed-off-by: Aaron Robson <aaron.robson@citrix.com>