CA-322672: Do not stop polling for RDP capabilities on VMs
These changes keep polling open console connections to check the status of RDP capabilities.
This is necessary when installing/uninstalling VM tools on VMs. Without this, changes in the `guest_metrics` aren't spotted. Consequently, the user is not shown the option to connect with RDP until they disconnect (or the console session is disconnected), even if the VM is capable.
That is because before this, the polling for RDP changes (as performed by `connectionPoller`) was only started if the VM had RDP capabilities when first connecting to it with the console.

Signed-off-by: Danilo Del Busso <Danilo.Del.Busso@citrix.com>