CA-378758: Ensure all hosts in pool hit `HOST_NOT_ENOUGH_FREE_MEMORY` before preventing migration
If pool any of the hosts in a pool hit the `HOST_NOT_ENOUGH_FREE_MEMORY` when calling `VM.assert_can_migrate`, XenCenter would prevent migration even if other hosts had enough memory to accommodate the VM.

Likely a result of the fact that other `VM.assert_can_migrate` failures are a consequence of pool-wide configurations. This is an exception.

Signed-off-by: Danilo Del Busso <danilo.delbusso@cloud.com>