Catch `HRESULT E_FAIL` exceptions for `IRDPClient.Connect` calls
The `Connect` method returns `E_FAIL` if it is called while the control is already connected or in the connecting state. This can be hit when a lot of connections are being opened at the same time, and it's there as a failsafe.

Also adds `IsAttemptingConnection` as a new field

Signed-off-by: Danilo Del Busso <danilo.delbusso@cloud.com>